+++ src/components/spec-renderer-toc/NodeGroupItem.vue
@@ -14,5 +14,4 @@ defineProps({
     required: true,
   },
 })
-
 </script>