+++ src/components/spec-renderer-toc/NodeItem.vue
@@ -1,20 +1,28 @@
 <template>
-  <li>
+  <li class="node-item">
     <a
+      :class="{ 'single-word': isSingleWord }"
       :href="`${basePath}${item.id}`"
       @click.prevent="selectItem(item.id)"
     >
       {{ item.title }}
+
+      <NodeItemBadge
+        v-if="item.type === 'http_operation'"
+        class="http-operation-badge"
+        :method="item.meta"
+      />
     </a>
   </li>
 </template>
 
 <script setup lang="ts">
-import { inject } from 'vue'
+import { computed, inject } from 'vue'
 import type { PropType } from 'vue'
 import type { TableOfContentsNode } from '../../stoplight/elements-core/components/Docs/types'
+import NodeItemBadge from './NodeItemBadge.vue'
 
-defineProps({
+const props = defineProps({
   item: {
     type: Object as PropType<TableOfContentsNode>,
     required: true,
@@ -31,4 +39,28 @@ const selectItem = (id: string): void => {
   emit('item-selected', id)
 }
 
+const isSingleWord = computed(() => !props.item.title?.trim()?.includes(' '))
 </script>
+
+<style lang="scss" scoped>
+@import '@/styles/mixins/mixins';
+
+.node-item {
+  list-style-type: none;
+
+  a {
+    @include toc-item;
+
+    &.single-word {
+      @include truncate;
+
+      display: block;
+    }
+
+    .http-operation-badge {
+      margin-left: var(--kui-space-auto, $kui-space-auto);
+      pointer-events: none;
+    }
+  }
+}
+</style>